Hole lot of debugging!

+++ Projects/Project5/sicc/population/infected.cpp
@@ -10,7 +10,7 @@ Infected::~Infected()
 
 Infected::Infected(int N, int I0, a_parameter *a, double b, double d, double dI)
 {
-    m_N = N; m_I0 = I0;
+    m_N = double(N); m_I0 = double(I0);
     m_b = b; m_d = d; m_dI = dI;
     m_a_param = a;
     current = double(m_I0);
